feat(puzzle): move offline vault setting to puzzles tab
- Vault choice lives on the puzzles home list, not inside a puzzle
- Shared picker helper;login-only like the offline queue setting
