+++ lib/src/view/puzzle/puzzle_tab_screen.dart
@@ -5,6 +5,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:lichess_mobile/src/model/auth/auth_controller.dart';
+import 'package:lichess_mobile/src/model/puzzle/offline_vault_prefs.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_angle.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_providers.dart';
@@ -21,6 +22,7 @@ import 'package:lichess_mobile/src/utils/screen.dart';
 import 'package:lichess_mobile/src/utils/string.dart';
 import 'package:lichess_mobile/src/view/account/account_menu.dart';
 import 'package:lichess_mobile/src/view/puzzle/dashboard_screen.dart';
+import 'package:lichess_mobile/src/view/puzzle/offline_vault_picker.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_history_screen.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_screen.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_themes_screen.dart';
@@ -272,6 +274,7 @@ class _PuzzleMenu extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final isOnline = ref.watch(isDeviceOnlineProvider);
     final authUser = ref.watch(authControllerProvider);
+    final vaultPrefs = ref.watch(offlineVaultPrefsProvider);
 
     return ListSection(
       hasLeading: true,
@@ -314,6 +317,13 @@ class _PuzzleMenu extends ConsumerWidget {
                 }
               : null,
         ),
+        if (authUser != null)
+          _PuzzleMenuListTile(
+            icon: Icons.download_for_offline_outlined,
+            title: 'Offline vault',
+            subtitle: 'Keep ${offlineVaultLabel(vaultPrefs)} on this phone.',
+            onTap: () => showOfflineVaultPicker(context, ref),
+          ),
         if (authUser != null) ...[
           _PuzzleMenuListTile(
             icon: Icons.assessment_outlined,
